requirements.txt
altgraph==0.17.4
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
macholib==1.16.3
packaging==24.0
pillow==10.3.0
pip==24.0
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.4
PyJWT==2.8.0
setuptools==65.5.0
sqlparse==0.5.0
typing==3.7.4.3
django-mptt==0.16.0
